Add a startup config

+++ src/main/java/me/lauriichan/minecraft/pluginbase/config/Configuration.java
@@ -240,7 +240,7 @@ public final class Configuration {
         if (object == null || !(object instanceof final Number value)) {
             return fallback;
         }
-        return value.doubleValue();
+        return value;
     }
 
     public boolean isConfiguration(final String pathUri) {
